requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.0
chromadb>=0.4.0
tqdm>=4.65.0
neo4j>=5.0.0
google-genai>=0.1.0
langfuse>=2.0.0
